database/requirements.txt
matplotlib
numpy
scipy
mysql-connector-python
#pymysql

